+++ lib/src/flutter/events/back_button_clicked_event.dart
@@ -1,8 +1,8 @@
 part of '../../../telegram_web_app.dart';
 
-/// Occurrs when the back button is pressed.
-/// [eventHandler] receives no parameters
+/// Bot API 6.1+ Occurrs when the back button is pressed.
+/// eventHandler receives no parameters.
 class BackButtonClickedEvent extends TelegramEvent {
-  BackButtonClickedEvent(Function() eventHandler)
+  BackButtonClickedEvent(void Function() eventHandler)
       : super(TelegramEventType.backButtonClicked, eventHandler);
 }