+++ lib/src/flutter/events/clipboard_text_received_event.dart
@@ -1,12 +1,16 @@
 part of '../../../telegram_web_app.dart';
 
-/// Occurs when the readTextFromClipboard method is called.
-/// [eventHandler] receives an object with the single field data
-/// containing text data from the clipboard. If the clipboard
-/// contains non-text data, the field data will be an empty string.
-/// If the Mini App has no access to the clipboard, the field data
-/// will be null.
+/// Bot API 6.4+ Occurrs when the readTextFromClipboard method is called.
+/// eventHandler receives an object with the single field data containing text data from
+/// the clipboard. If the clipboard contains non-text data, the field data will be an empty
+/// string. If the Mini App has no access to the clipboard, the field data will be null.
 class ClipboardTextReceivedEvent extends TelegramEvent {
-  ClipboardTextReceivedEvent(Function(String) eventHandler)
+  ClipboardTextReceivedEvent(void Function(ClipboardTextReceivedPayload payload) eventHandler)
       : super(TelegramEventType.clipboardTextReceived, eventHandler);
+}
+
+class ClipboardTextReceivedPayload {
+  final String? data;
+
+  ClipboardTextReceivedPayload(this.data);
 }