+++ lib/src/flutter/events/contact_requested_event.dart
@@ -1,11 +1,17 @@
 part of '../../../telegram_web_app.dart';
 
-/// Occurs when the user's phone number was requested.
-/// [eventHandler] receives an object with the single field
-/// status containing one of the statuses:
+/// Bot API 6.9+ Occurrs when the user's phone number was requested.
+/// eventHandler receives an object with the single field status containing one of the
+/// statuses:
 /// - sent – user shared their phone number with the bot,
 /// - cancelled – user declined this request.
 class ContactRequestedEvent extends TelegramEvent {
-  ContactRequestedEvent(Function(String status) eventHandler)
+  ContactRequestedEvent(void Function(ContactRequestedPayload payload) eventHandler)
       : super(TelegramEventType.contactRequested, eventHandler);
+}
+
+class ContactRequestedPayload {
+  final String status;
+
+  ContactRequestedPayload(this.status);
 }