+++ lib/src/flutter/events/popup_closed_event.dart
@@ -1,11 +1,15 @@
 part of '../../../telegram_web_app.dart';
 
-/// Occurs when the opened popup is closed.
-// [eventHandler] receives an object with the
-// single field [button_id] – the value of the
-// field id of the pressed button. If no buttons
-// were pressed, the field button_id will be null.
+/// Bot API 6.2+ Occurrs when the opened popup is closed.
+/// eventHandler receives an object with the single field button_id – the value of the field
+/// id of the pressed button. If no buttons were pressed, the field button_id will be null.
 class PopupClosedEvent extends TelegramEvent {
-  PopupClosedEvent(Function(String? buttonId) eventHandler)
+  PopupClosedEvent(void Function(PopupClosedPayload payload) eventHandler)
       : super(TelegramEventType.popupClosed, eventHandler);
+}
+
+class PopupClosedPayload {
+  final String? buttonId;
+
+  PopupClosedPayload(this.buttonId);
 }