+++ lib/src/flutter/events/write_access_requested_event.dart
@@ -1,11 +1,17 @@
 part of '../../../telegram_web_app.dart';
 
-/// Occurs when the write permission was requested.
-/// [eventHandler] receives an object with the single
-/// field status containing one of the statuses:
+/// Bot API 6.9+ Occurs when the write permission was requested.
+/// eventHandler receives an object with the single field status containing one of the
+/// statuses:
 /// - allowed – user granted write permission to the bot,
 /// - cancelled – user declined this request.
 class WriteAccessRequestedEvent extends TelegramEvent {
-  WriteAccessRequestedEvent(Function(WriteAccessRequestedEventResult result) eventHandler)
+  WriteAccessRequestedEvent(void Function(WriteAccessRequestedPayload result) eventHandler)
       : super(TelegramEventType.writeAccessRequested, eventHandler);
+}
+
+class WriteAccessRequestedPayload {
+  final String status;
+
+  WriteAccessRequestedPayload(this.status);
 }